Use a tagged relase of Mynewt-nimble

+++ src/libs/mynewt-nimble/nimble/controller/src/ble_ll_conn_priv.h
@@ -120,7 +120,7 @@ struct hci_create_conn
 void ble_ll_conn_sm_new(struct ble_ll_conn_sm *connsm);
 void ble_ll_conn_end(struct ble_ll_conn_sm *connsm, uint8_t ble_err);
 void ble_ll_conn_enqueue_pkt(struct ble_ll_conn_sm *connsm, struct os_mbuf *om,
-                             uint8_t hdr_byte, uint16_t length);
+                             uint8_t hdr_byte, uint8_t length);
 struct ble_ll_conn_sm *ble_ll_conn_sm_get(void);
 void ble_ll_conn_master_init(struct ble_ll_conn_sm *connsm,
                              struct hci_create_conn *hcc);
